HBASE-2054  memstore size 0 is >= than blocking -2.0g size HBASE-2057  Cluster won't stop
git-svn-id: https://svn.apache.org/repos/asf/hadoop/hbase/branches/0.20@892391 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/hadoop/hbase/HTableDescriptor.java
@@ -92,9 +92,9 @@ ISerializable {
 
   public static final boolean DEFAULT_READONLY = false;
 
-  public static final int DEFAULT_MEMSTORE_FLUSH_SIZE = 1024*1024*64;
+  public static final long DEFAULT_MEMSTORE_FLUSH_SIZE = 1024*1024*64L;
   
-  public static final int DEFAULT_MAX_FILESIZE = 1024*1024*256;
+  public static final long DEFAULT_MAX_FILESIZE = 1024*1024*256L;
     
   private volatile Boolean meta = null;
   private volatile Boolean root = null;
@@ -402,7 +402,7 @@ ISerializable {
   /**
    * @return memory cache flush size for each hregion
    */
-  public int getMemStoreFlushSize() {
+  public long getMemStoreFlushSize() {
     byte [] value = getValue(MEMSTORE_FLUSHSIZE_KEY);
     if (value != null)
       return Integer.valueOf(Bytes.toString(value)).intValue();
